moved signUpValidator to middlewares folder and deleted try/catch block in usersController signUp method
